Time: 7 ms (10.68%), Space: 5.9 MB (60.02%) - LeetHub

+++ 2224-minimum-number-of-operations-to-convert-time/2224-minimum-number-of-operations-to-convert-time.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     int convertTime(string current, string correct) {
-        long long hour1 = stoll(current.substr(0, 2)), min1 = stoll(current.substr(3, 5));
-        long long hour2 = stoll(correct.substr(0, 2)), min2 = stoll(correct.substr(3, 5));
+        int hour1 = stoll(current.substr(0, 2)), min1 = stoll(current.substr(3, 5));
+        int hour2 = stoll(correct.substr(0, 2)), min2 = stoll(correct.substr(3, 5));
         int ret = 0;
         if(min1 > min2) {
             --hour2;
